+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/sync/SyncHistoryTask.kt
@@ -14,7 +14,6 @@ import au.com.shiftyjelly.pocketcasts.models.to.HistorySyncRequest
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.HistoryManager
-import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
 import au.com.shiftyjelly.pocketcasts.utils.extensions.switchInvalidForNow
 import au.com.shiftyjelly.pocketcasts.utils.extensions.toIsoString
 import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
@@ -27,12 +26,11 @@ private const val TAG = "SyncHistoryTask"
 
 @HiltWorker
 class SyncHistoryTask @AssistedInject constructor(
-    @Assisted val context: Context,
+    @Assisted context: Context,
     @Assisted params: WorkerParameters,
-    var episodeManager: EpisodeManager,
-    var syncManager: SyncManager,
-    var podcastManager: PodcastManager,
-    var settings: Settings,
+    private val episodeManager: EpisodeManager,
+    private val syncManager: SyncManager,
+    private val settings: Settings,
     private val historyManager: HistoryManager,
 ) : CoroutineWorker(context, params) {