+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 7.83
 -----
+*   New Features
+    *   Add support for updating podcast feed
+        ([#3569](https://github.com/Automattic/pocket-casts-android/pull/3569))
 *   Bug Fixes
     *   Fix missing episodes with unfollowed podcasts
         ([#3546](https://github.com/Automattic/pocket-casts-android/pull/3546))
@@ -10,6 +13,8 @@
 *   Updates
     *   Filter out chapters that do not belong in table of contents. See [the specification](https://github.com/Podcastindex-org/podcast-namespace/blob/main/chapters/jsonChapters.md) for more details.
         ([#3556](https://github.com/Automattic/pocket-casts-android/pull/3556))
+    *   Add OPML import from URL to the onboarding process.
+        ([#3584](https://github.com/Automattic/pocket-casts-android/pull/3584))
 
 7.82
 -----